tests(ReviseTone): Limited Cypress tests to ensure Revise Tone shows once
The recent change I00d12f0094 enabled the use of TestKitchen by default
for all of GrowthExperiments. This makes sense, but it also means that
in our tests we need to account for that. However, MetricsPlatform is
not available in the quibble-with-gated-extensions-selenium-php83 job.

While that is something that should be fixed in general, for now this
just disables the use of TestKitchen during Cypress tests.

For future reference, a version of this change that would have made use
of TestKitchen to override this group is available on Gerrit in
PatchSet 15.

Change-Id: Ib618396367da82f8536029d2c876de71e290ef31