MentorQuestionPoster: Refactor so ::construct works without a mentor
On a related gerrit change, change-id I102554c9e55eae9917f527daa6d312170cb3bce0, Urbanecm had suggested refactoring QuestionPoster to have a method that tells if the question can be asked. Previously, this wasn't possible because a MentorQuestionPoster without a mentor would throw an exception during construction. Due to the lazy loading in this patch, a MentorQuestionPoster without a mentor won't throw an error until submit() is called.

Bug: T386567
Change-Id: I3c1c5ccdae75218f868a83eb3230e33c98e714e3