UserImpact: stop using pre-computed impact in the user impact job
When increasing wgGEUserImpactMaxEdits to reach a higher bound
of edits data, the number of edited articles will grow for many
users, making the page view data size bigger and the overall user
impact blob size.

Avoid reusing the pre-computed impact in the job context to prevent
hitting the job param limit. This could lead to some data
discrepancies in the UserImpactHandler API GET code path which is only
used for Wikipedia Mobile App. The discrepancy could be mitigated
by passing an array of priority articles in the handler computation
but that is left for a follow-up once we can check it exists indeed.

Bug: T398500
Change-Id: I877ffff5e07e29e639ced65da4f82e11538178f9