fix(ReviseTone): don't trigger an exception on missing data
In \GrowthExperiments\HomepageHooks::onBeforePageDisplay we are
requesting the data for a structured task from the respective
data-provider, in this case, ApiReviseToneRecommendationProvider.

If that provider returns a Status, then we log that via Util::logStatus.
However, that also triggers an exception to be logged if that Status is
not OK. That is not something we want here, because a page having
unexpectedly is something that can happen (rarely) in normal production.

This is also already documented in the RecommendationProvider interface:

> @return Recommendation|StatusValue The recommendation, or a StatusValue on error.
>	    The StatusValue's OK flag will determine whether this is an unexpected error that
>	    should be sent to the production error logs, or something that can happen under normal
>	    circumstances (e.g. the given article simply not having any recommendations).

Bug: T411855
Change-Id: I5fc7ac2e0cc994053418b01b30fb28602f93b439